Move validated loading into loader
